classic_api/frontend/requirements.txt
streamlit
httpx
langchain
langchain-openai
python-dotenv
markdown